requirements.txt
python-pptx>=0.6.21
pandas>=2.0.0
openpyxl>=3.1.0
mcp>=0.1.0
